pyproject.toml
[tool.poetry]
name = "use-drf"
version = "0.1.0"
description = ""
authors = ["stas <stas.skibickiy@gmail.com>"]
readme = "README.md"
packages = [{include = "use_drf"}]

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.2"
generics = "^6.0.0"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
